Algorythm for retrieving page of search

+++ src/main/java/de/fiz/oai/backend/models/SearchResult.java
@@ -3,17 +3,17 @@ package de.fiz.oai.backend.models;
 import java.util.List;
 
 public class SearchResult<E> {
-  int total;
+	long total;
   int offset;
   int size;
   
   List<E> data;
 
-  public int getTotal() {
+  public long getTotal() {
     return total;
   }
 
-  public void setTotal(int total) {
+  public void setTotal(long total) {
     this.total = total;
   }
